Command <-> BP, Feet Offset
BP doesn't accept pointers to UStructs, which Commands are.
CommandDrivenController is going to need functions that combine
ReceiveCommand with the various Command constructors as a result.
Actors' locations are at the center of the actor. Not helpful for
navigation (which 99% of the time uses the floor). GetFeetLocation
should be used instead.